Add new test files for `tox` to work:
Add the `test-requirements.in` and compiled `test-requirements.txt`
requirements files that operate like the non-test
`requirements.in`/`requirements.txt` files. These files are used to
define the testing dependencies used by `tox`.

Modify the `tox.ini` configuration to reference the existence of
`test-requirements.txt` so that the dependencies on tests are met.

`requirements.in`/`requirements.txt` has been modified to migrate some
dependencies into `test-requirements.*`.