Create `tox` workflow to run on project: (#9)
Update `tox` to successfully run the `tests`/`pep8`/`mypy` actions on the 
codebase.

This required the introduction of a `test-requirements.in` and compiled 
`test-requirements.txt` which operates the same as the existing
`requirements.in`/`requirements.txt` files.

Update the `tox.ini` configuration file to use the `test-requirements.txt` file,
and also update the placeholder `[CODEDIR]` to the `qcp/` dir (for now?).

Some minor tweaks to the `setup.py` file were required for it to pass the now
working `mypy` checks.

Finally, generate a workflow job using https://github.com/medmunds/tox-gh-matrix to 
allow running of `tox` actions as workflow steps on this repo.

Currently the `tox / tests` action reports a failure, this is expected as `pytest` exits with
code 5 when no test files are found, which results in the action failing. This
will be resolved once test code is implemented.