Add placeholder test file: (#15)
Add a simple `test_main.py` file in `tests/` that verifies the `main.py` code
exits with no error. This is a no brainer as currently `main.py` just calls
`pass`. However, a test file is required to prevent pytest from exiting with
error code 5, and marking the `tests` workflow action as failing. 

This PR should resolve that.