Implement the `tensor_product()` method:
Method takes in two matrices and computes the tensor product as a larger
matrix, and returns that matrix.

The tensor product of two matrices A, B is defined as:

A x B = [ A00 * B, A01 * B, ... A0n * B ]
	[ A10 * B, A11 * B, ...	A1n * B ]
	[ ...			...	]
	[ An0 * B, An1 * B, ... Ann * B ]

Resulting in an (mp)*(np) matrix, if A is m*n & B is n*p.

The code works by iterating over the indices of the full mp*np matrix,
and populating the entries. To calculate the factor of A to use, the
current indices are floor divided by the width of A (= n). This factor
is then multiplied against the corresponding B entry, which is
calculated by taking the modulus of the indices.

Calculting the tensor product in this way results in an O(n^2)
calculation.