Implement tensor product code: (#23)
Implement the `tensor_product()` method:

Method takes in two matrices and computes the tensor product as a larger
matrix, and returns that matrix.

The tensor product of two matrices A, B is defined as:

```
A x B = [ A00 * B, A01 * B, ... A0n * B ]
        [ A10 * B, A11 * B, ... A1n * B ]
        [ ...                   ...     ]
        [ Am0 * B, An1 * B, ... Amn * B ]
```

Resulting in an `(mp)*(nq)` matrix, if A is `m*n` & B is `p*q`.

The code works by iterating over the indices of the full `mp*np` matrix,
and populating the entries. To calculate the factor of A to use, the
current indices are floor divided by the length/width of A (= n & m). 
This factor is then multiplied against the corresponding B entry, which 
is calculated by taking the modulus of the indices against it's dimensions 
(p & q).

To make reviewing the code changes easier, the following smaller changes
have been spun out to their own PRs:
* https://github.com/Tiernan8r/quantum_computing_project/pull/19
* https://github.com/Tiernan8r/quantum_computing_project/pull/20
* https://github.com/Tiernan8r/quantum_computing_project/pull/21
* https://github.com/Tiernan8r/quantum_computing_project/pull/22