Sparse Matrix Implementation: (#27)
Add an implementation of a `SparseMatrix` in python.

The sparse matrix works by only keeping a record of non-zero (within 1e-09 of 0)
values as a mapping of indices in a python dict.

The python dict type was chosen to represent the data on the backend due to it's
optimised lookup speeds, low memory usage, and indexing allows the easy
retrieval of values.

This PR requires both:
* https://github.com/Tiernan8r/quantum_computing_project/pull/23
* https://github.com/Tiernan8r/quantum_computing_project/pull/24
to land before it is ready to merge, as it builds on the implementations
introduced in these PRs.

Some peripheral changes to configuration and test files were required to ease my
introduction of tests for the SparseMatrix. According to `tox`, test coverage
currently stands at 97% as a result.

The SparseMatrix fully implements all the stubs present in `Matrix`, some new
stubs have been introduced that I found convenient while implementing this
class.