Enable Dependabot on this repo:
Dependabot is a tool provided by Github that tracks the dependencies of
your codebase, and proposes updates to the dependencies whenever a new
version becomes available.

Enable this functionality for the `pip` packages and github workflows in
this repo.

Also update the `test-requirements.in` manifest to remove some unused
dependencies. `tox -e update-pins` was re-run to recompile
`requirements.txt` and `test-requirements.txt`.