Patch error in `SparseMatrix` dot product calculation:
Fix an error that myself and Finn came across in how the `SparseMatrix`
calculates the dot product. The resultant matrix's row/column dimensions
were incorrect after the calculation when taking the dot product of
non-square matrices.

This should hopefully be patched now, and a new test exists that should
catch the error arising again.