Fix edge case in Matrix dot producting: (#53)
Fix index error when calculating the dot product between row and column vectors
due to missing indices in the dictionaries of the underlying `SparseMatrix`
objects.

This error was occuring in two places, firstly when the `transpose()` was being
calculated, the dictionary rows were only populated if they had non-zero
entries, while the general `SparseMatrix` code always assumed the zero rows were
represented as empty dicts.

Secondly, since most of the tests are implemented using square matrices, there
was an error in which dimension to use in the new matrix when calculating the
dot product, previously the number of rows of the new matrix was set using the
number of rows of the current matrix, when it should have been using the number
of columns. This has been resolved.

A new test has been added to verify the code works as expected.