Optimise SparseMatrix `_dot()` method: (#51)
Introduce a new submethod to `_dot()` in `SparseMatrix` that calculates the dot
product between two `SparseMatrix` objects.

This allows us to take advantage of the sparseness of the `other` object, and
only calculate multiplications between non-zero entries. For highly sparse
matrices, this should hopefully result in a speed-up of the calculations.