Basic CLI: (#55)
Implement a basic CLI (Command Line Interface) for the program. 

Basic usage of the CLI can be optained by running the following command:
```console
$ python -m qcp.main --help
USAGE:
/home/tiernan8r/quantum_computing_project/qcp/main.py [FLAGS] nqbits

ARGS:
    nqbits          The number of qbit states to simulate
FLAGS:
    -t/--target     The target state, defaults to 5
    -h/--help       Display this prompt
```

The `main.py` file has been modified to make this possible. When `main()` is
called, the arguments passed in via the terminal are parsed and filtered by the
`parse_cli()` method, which returns the arguments we are interested in.
The code for Grover's Algorithm now runs in a `compute()` method, which takes in
the required parameters and runs the Algorithm like before. Importantly, the
`compute()` and `main()` method need to be separate if we want to also implement
a GUI later.

As a drive-by, modify the requirements files to move the testing related
requirements of `tox`/`pip-tools` to the `test-requirements.in` file, and
recompile. This higlights the fact that our Quantum Computer simulator is
implemented in pure Python (yay!)