UI Foundation: (#59)
Implement the foundations for the UI of Grover's Algorithm.

Currently the UI interactions are not tied into the backend of the Quantum
Computer Simulator implementation, so the UI is not currently functional as an
interface for our simulator.

The UI is implemented in Qt [qt.io](qt.io) using the [PySide6](https://pypi.org/project/PySide6/) library. 
This has been added as a requirement in `requirements.txt`.

The UI is designed and laid out in the `form.ui` file, and then loaded at
runtime using the `main_window.py` file. The behaviour of sections of the UI are
implemented in files under the `components/` module, where button interactions,
filling textboxes and the likes are given relevant UI behaviours instead of
being static.

When it comes to tying in the UI to the actual simulator, this can be done in
the `simulator_component.py` file, where the skeleton of running the simulation
in a separate thread is implemented.

The UI can be run via the following command from the top level of our project
directory:
```console
$ python -m qcp.ui.main
```

This will (hopefully) bring up the following interface:

![Gabháil scáileáin (11)](https://user-images.githubusercontent.com/9886002/158068793-a53f15fd-5e1f-4fd4-83e3-4dddefe6a525.png)

The "Search" box can either be populated manually, or by clicking the "browse"
button, a file dialog to open a file of the type `.txt`/`.json`/`.csv`/`.yml`
can be opened and it's content imported into the text box.

![Gabháil scáileáin (14)](https://user-images.githubusercontent.com/9886002/158068809-d8d03268-3f33-46bf-a050-51301e47b483.png)

The "Target" field must then be manually filled.

When the "Search" button is hit, a progress ticker will appear, along with a
"cancel" button, which will stay around until the simulation is complete and the
results calculated (not implemented)

The two "count" fields will then increment to the number of iterations a
classical computer would take, versus a quantum computer, for the same search
in the data (not implemented).

When the data is calculated, an embedded matplotlib plot will appear on the
right, showing the calculated probabilities of each state (not implemented),
currently a placeholder graph will appear. There is also currently a bug in how
the graph is displayed, it does not auto resize the window to correctly display
the graph, it must be resized manually.

![Gabháil scáileáin (13)](https://user-images.githubusercontent.com/9886002/158068817-7a6de917-1670-426c-8c49-c4681adc63ac.png)