Update `tox.ini`:
Update the `tox.ini` configuration to run the tests on our quantum simulator
code, and tests on the UI in two separate steps. This ensures that the coverage
reports on our main simulation code is reported correctly, as the UI code does
not have test code, and is bringing down the averages on the coverage reports
drastically.