Fix error in C-PHASE gate:
Fix the implementation of the control phase gate, so that the bitmask sets the
phase flip for the target bit correctly based off of the control bits.

This bit masking is the same logic that is being used in `control_z()`.
To save us duplicating code, create a generic private method
`_generic_control()` that sets the matrix index to the given value based off of
our control and target inputs.

Implement this new generic matrix for the `control_z()` and `control_phase()`.
The `control_z()` tests do not need to be modified for this change, but update
the `control_phase()` tests to verify we get the expected behaviour.