Docstring everything: (#67)
Add docstrings to all methods in the code, so that the Sphinx docs that are
generated will document all the code in our repository.

Where docstrings existed already, they are fixed where appropriate so they are
displayed properly in HTML.

Update the Sphinx configuration so that the README document is automatically
read into the index.html file, so that the content is reproduced in the HTML
format too.