Modify the UI input:
Change the UI usage so that the first text box sets the number of qbits to
simulate in our system, and the second one sets the target qbit.

The code now runs the simulation of Grover's Algorithm and plots the result on
the UI.

If there is an error with the input, say we don't provide an integer for num
qbits, then a red error text box will appear with the error message.

Remove the browse file dialog as it is not needed for this iteration.