Use BibLatex for citations: (#78)
Change the report to use BibLatex for citations:

Add a new file `references.bib` containg all the current references that
were defined at the end of the `main.tex` file.

Convert all the `\hyperlink` commands in the report to use `\cite{}`
with the relevant citation. LaTeX will figure out the bibliography and
citation numbers after that.