Make the `main.py` file runnable directly: (#87)
Run `chmod +x qcp/main.py` to give the `main.py` file executable
permissions. This means we can now run the CLI by the following command:

```console
$ ./qcp/main.py
```

Add a simple check at the start of `main.py` to ensure the 'qcp' module
is in the system path, otherwise the code can fail with a ModuleNotFound
error.