Streamline the Algorithms code: (#90)
Streamline the different implementations of quantum algorithms by the usage of a
generalised algorithm interface.

Move the three algorithm implementations to a new submodule `algorithms`.

Add a new `abstract_algorithm.py` class that is of type `ABC`. It defines some
default behaviours of our algorithms, and can reduce the amount of code
duplication in the codebase.

Tweak the `Grover`/`PhaseEstimation`/`Sudoku` implementations to now inherit
from this interface, and remove duplicated code where needed. Some lines had to
be re-organised to ensure the inheritance method calling didn't error out with
missing variables.