Fix the CLI usage, and prettify:
Fix a bug in the CLi usage where the sudoku algorithm could not be run and
instead the help message was being displayed because no arguments were provided.

Prettify the output from each algorithm type.

Show a table of probabilities along with the measured result for Grover and
Phase Estimation.

For sudoku, should a table of the sudoku, with the probability of that being
right.