Update & Finalise GUI: (#95)
Unfortunately a chonker of a PR :(

Update the GUI to work for all algorithm types

The GUI can now be run by the simplified command:
```console
$ ./qcp/gui/main.py
```

Rename the `qcp/ui` directory to `qcp/gui` as this makes more sense to me.

Generalise the components implementation.

The `abstract_component.py`/`graph_component.py` have been kept in the top-level
`qcp/gui/components/` submodule.

All pre-existing components related to Grover's Algorithm have been moved to a
`qcp/gui/components/grovers` submodule.

With equivalents for `phase_estimation`/`sudoku` (I'll get to those...)

A new `combo_box_component.py` has been added, which controls the interaction
with the algorithm drop down menu within the UI, and toggles the UI elements to
show depending on the choice.

The interaction with the progress bar has been generalised to a
`progress_bar_component.py` component, with all button component classes that
used to interact with it directly now inheriting from this class to reduce
duplication.

Similarly for the `SimulateQuantumComputerThread`, this has been generalised and
abstracted to now take any number of arguments as a tuple, as well as the
constructor for the algorithm to simulate, and returns a general tuple of the
results, it is up to the implementations of this thread to make sure the I/O
behaves properly.

The three algorithm types have been implemented into the UI.

**Grover's Algorithm:**

![gui_overview_grovers](https://user-images.githubusercontent.com/9886002/159334624-79effb5f-9a7f-4737-8e64-dc29f57f0113.png)

For Grover's Algorithm, not much has changed. The input boxes are now QSpinBox
components that allows for greater control of the `nqbits` and `target` state
values, and enforces natively that they are integers.

![gui_calculation_grovers](https://user-images.githubusercontent.com/9886002/159334781-90eca49a-7873-452d-8f3c-b2dbdd5bd8c8.png)

**Phase Estimation Algorithm:**

![gui_overview_phase_estimation](https://user-images.githubusercontent.com/9886002/159334865-f676337b-99b4-402e-910f-1aa29c41e950.png)

Input for the number of qbits is also provided as a spin box.

The input of the unitary matrix and the eigenvector matrix are handled via
QTableWidget objects. Where they are resized based off of the number of qbits to
simulate.

![gui_calculation_phase_estimation](https://user-images.githubusercontent.com/9886002/159334912-aa6814fd-9bc3-43ff-b0ce-8fcaed3f6db7.png)

**Sudoku Algorithm:**

![gui_overview_sudoku](https://user-images.githubusercontent.com/9886002/159334960-f5943515-c425-4f83-8f81-5a7358d72aad.png)

It takes no input so was easy to implement, the start/cancel buttons are hooked
up to the backend and progress bar as expected.

Results for the Sudoku algorithm require special handling as there are two sets
of output, the usual values that are plottable.

And a new output, of the calculated sudoku entries which are displayed in a
table, along with the given probability of that scenario occuring, which is
visualised using another progress bar widget which natively displays as
percentages.

![gui_calculation_sudoku](https://user-images.githubusercontent.com/9886002/159335002-330336fc-c116-4baf-87b9-c87d9202e2f4.png)