Redo the Phase Estimation GUI/CLI: (#100)
Redo how the Phase Estimation frontend is interacted with:

Update the phase estimation GUI based off of incredibly useful comments from @nys1998 on PR
https://github.com/Tiernan8r/quantum_computing_project/pull/95.

As an addendum on the GUI updating, all algorithm GUI elements now show the measured state
and the probability of measuring that state by default.

**GUI**

The updated GUI now looks like this:

![gui_overview_phase_estimation](https://user-images.githubusercontent.com/9886002/159533560-081a63df-be59-420f-974e-fa06ec37b9f2.png)

As before, the user can input the number of qbits to use in the simulation.

The user can now select some pre-defined unitary matrices to use, from a drop down list, 
this list of options is auto-populated from the `UnitaryMatrices` enum, which now lives in
`phase_estimation_unitary_matrices.py`.

When the user selects a unitary matrix, the table showing it's values will update.

If the user selects the `phase shift` unitary matrix, two new fields `phi1`/`phi2` will appear,
allowing the user to set the phase to use in this unitary matrix. The matrix will update as the
values are changed.

![gui_overview_phase_estimation_2](https://user-images.githubusercontent.com/9886002/159533615-43e4f632-3c67-44cf-b04d-eb87df542254.png)

The two unitary matrix options `hadamard`/`phase_shift` work with different eigenvectors, these
eigenvectors can also be chosed from a drop down list, which is also autopopulated.

As before, the table below the drop down visually updates to show the eigenvector chosen.

At the bottom of the GUI is a new section which allows the user to auto-set the number of
qbits to use in the simulation, based off of their desired decimal precision, and acceptable
error rate. hitting the "Determine" button will auto-set the "number of qbits" field.

![gui_calculation_phase_estimation](https://user-images.githubusercontent.com/9886002/159533656-fc7de37c-8495-4ede-90e5-46bfd19d7279.png)

**CLI:**

Usage of the CLI has been updated with similar functionality:

The help message now looks like this:

```console
$ ./qcp/main.py -h
USAGE:
./qcp/main.py [FLAGS] nqbits
FLAGS:
    -a/--algorithm  The quantum algorithm to simulate, can be one of:
                        g   = Grover's Algorithm
                        pe  = Phase Estimation
                        s   = Toy Sudoku solver
                    Defaults to 'g' if unset
    -h/--help       Display this prompt
    -g/--gui      Display the GUI (if supported).

The CLI options vary by choice of algorithm:

GROVERS:
    USAGE:
    ./qcp/main.py --algorithm g [FLAGS] nqbits
    ARGS:
        nqbits          The number of qbit states to simulate, must be >= 2.
    FLAGS:
        -t/--target     The target state, defaults to 0

PHASE ESTIMATION:
    USAGE:
    ./qcp/main.py --algorithm pe [FLAGS] nqbits
    ARGS:
        nqbits          The number of qbit states to simulate, must be >= 2.
    FLAGS:
        -p/--phase      The phase to use for the Phase Shift Gate (if using).
                        defaults to 0.25.
        -u/--unitary    The choice of unitary gate to simulate with, can be
                        one of the following:
                            * hadamard = Hadamard Gate
                            * phase_shift = Phase Shift Gate (requires the -p flag)
                        Defaults to 'hadamard' if unset
        -e/--eigen      The choice of eigenvector to use in the simulation.
                        IF using the HADAMARD gate, can be one of:
                            * +
                            * -
                            Default is +
                        If using the PHASE SHIFT gate, can be one of:
                            * 0
                            * 1
                            Default is 0
SUDOKU:
    USAGE:
    ./qcp/main.py --algorithm s
    ARGS:
        None
    FLAGS:
        None
```

When running the phase estimation algorithm, the user can now input
the choice of unitary matrix, and based off of that choice, a choice
of eigenvectors. The symbols chosen to represent the eigenvectors
are derived from the `|0>`/`|1>`/`|+>`/`|->` notation, but in the
terminal, the `|` & `>` symbols are pipe characters, so cannot be typed
easily, so are dropped in the CLI input.