Change tox test coverage to not report on GUI code:
This generates a table of test coverage that is related to code
we have actually written tests for, rather than the GUI code which
is large and has no test code.

More of a quality of life PR, just results in the coverage table reported
by tox being more realistic in it's test coverage.